requirements.txt
nympy
matplotlib
numpy-hilbert-curve